Passage des niveaux de gris en double dans ImagePoint

+++ src/helper/TestingImageReader.java
@@ -83,7 +83,7 @@ public class TestingImageReader{
 		// System.out.println( label );
 		labelOffset++;
 
-		int[] img = new int[width * height];
+		double[] img = new double[width * height];
 
 		for(int i = 0; i < width; i++)
 			for(int j = 0; j < height; j++){
